+++ src/allotropy/parsers/unchained_labs_lunatic_stunner/unchained_labs_lunatic_stunner_parser.py
@@ -9,17 +9,17 @@ from allotropy.allotrope.schema_mappers.adm.plate_reader.rec._2025._03.plate_rea
 )
 from allotropy.named_file_contents import NamedFileContents
 from allotropy.parsers.release_state import ReleaseState
-from allotropy.parsers.unchained_labs_lunatic.unchained_labs_lunatic_reader import (
+from allotropy.parsers.unchained_labs_lunatic_stunner.unchained_labs_lunatic_reader import (
     UnchainedLabsLunaticReader,
 )
-from allotropy.parsers.unchained_labs_lunatic.unchained_labs_lunatic_structure import (
+from allotropy.parsers.unchained_labs_lunatic_stunner.unchained_labs_lunatic_structure import (
     create_measurement_groups,
     create_metadata,
 )
 from allotropy.parsers.vendor_parser import VendorParser
 
 
-class UnchainedLabsLunaticParser(VendorParser[Data, Model]):
+class UnchainedLabsLunaticStunnerParser(VendorParser[Data, Model]):
     DISPLAY_NAME = "Unchained Labs Lunatic & Stunner"
     RELEASE_STATE = ReleaseState.RECOMMENDED
     SUPPORTED_EXTENSIONS = UnchainedLabsLunaticReader.SUPPORTED_EXTENSIONS
