+++ tests/parsers/unchained_labs_lunatic_stunner/testdata/2024-07-12_CW288_Plate1.xlsx
