+++ tests/parsers/unchained_labs_lunatic_stunner/testdata/2024-07-16_CW288_Plate2.xlsx
