+++ tests/parsers/unchained_labs_lunatic_stunner/testdata/Demo_A260_dsDNA_Data.json
@@ -4142,7 +4142,7 @@
         "data system document": {
             "ASM file identifier": "Demo_A260_dsDNA_Data.json",
             "data system instance identifier": "N/A",
-            "UNC path": "tests/parsers/unchained_labs_lunatic/testdata/Demo_A260_dsDNA_Data.csv",
+            "UNC path": "tests/parsers/unchained_labs_lunatic_stunner/testdata/Demo_A260_dsDNA_Data.csv",
             "file name": "Demo_A260_dsDNA_Data.csv",
             "ASM converter name": "allotropy_unchained_labs_lunatic_&_stunner",
             "ASM converter version": "0.1.102",
