+++ tests/parsers/unchained_labs_lunatic_stunner/testdata/Demo_A280_Protein.json
@@ -6486,7 +6486,7 @@
         "data system document": {
             "ASM file identifier": "Demo_A280_Protein.json",
             "data system instance identifier": "N/A",
-            "UNC path": "tests/parsers/unchained_labs_lunatic/testdata/Demo_A280_Protein.csv",
+            "UNC path": "tests/parsers/unchained_labs_lunatic_stunner/testdata/Demo_A280_Protein.csv",
             "file name": "Demo_A280_Protein.csv",
             "ASM converter name": "allotropy_unchained_labs_lunatic_&_stunner",
             "ASM converter version": "0.1.102",
