+++ tests/parsers/unchained_labs_lunatic_stunner/testdata/Example_Lunatic_Plate_Reader_csv_no_header.json
@@ -212,7 +212,7 @@
         "data system document": {
             "ASM file identifier": "Example_Lunatic_Plate_Reader_csv_no_header.json",
             "data system instance identifier": "N/A",
-            "UNC path": "tests/parsers/unchained_labs_lunatic/testdata/Example_Lunatic_Plate_Reader_csv_no_header.csv",
+            "UNC path": "tests/parsers/unchained_labs_lunatic_stunner/testdata/Example_Lunatic_Plate_Reader_csv_no_header.csv",
             "file name": "Example_Lunatic_Plate_Reader_csv_no_header.csv",
             "ASM converter name": "allotropy_unchained_labs_lunatic_&_stunner",
             "ASM converter version": "0.1.102",
