+++ tests/parsers/unchained_labs_lunatic_stunner/testdata/Example_Lunatic_Plate_Reader_xlsx_with_header.json
@@ -11,6 +11,7 @@
                                     {
                                         "device type": "plate reader",
                                         "detection type": "Absorbance",
+                                        "firmware version": "8.3.0.305",
                                         "detector wavelength setting": {
                                             "value": 260.0,
                                             "unit": "nm"
@@ -25,6 +26,10 @@
                                 "batch identifier": "GROUP_1",
                                 "well plate identifier": "PLATE_1",
                                 "custom information document": {
+                                    "plate type": "Lunatic Plate",
+                                    "nr of plates": "1",
+                                    "blanks": "Sample group 'GROUP_1': On Plate/Chip:  (2x)",
+                                    "plate description": "- Lunatic_test_BE_31JUL2024",
                                     "analyte": "N/A"
                                 }
                             },
@@ -44,7 +49,8 @@
                                         "error feature": "Background (A260)"
                                     }
                                 ]
-                            }
+                            },
+                            "experimental data identifier": "BENCHLING_TEST"
                         }
                     ],
                     "measurement time": "2024-01-08T10:15:59+00:00",
@@ -53,7 +59,8 @@
                         "unit": "#"
                     },
                     "container type": "well plate"
-                }
+                },
+                "analyst": "ImmuneMed"
             },
             {
                 "measurement aggregate document": {
@@ -64,6 +71,7 @@
                                     {
                                         "device type": "plate reader",
                                         "detection type": "Absorbance",
+                                        "firmware version": "8.3.0.305",
                                         "detector wavelength setting": {
                                             "value": 260.0,
                                             "unit": "nm"
@@ -78,6 +86,10 @@
                                 "batch identifier": "GROUP_1",
                                 "well plate identifier": "PLATE_1",
                                 "custom information document": {
+                                    "plate type": "Lunatic Plate",
+                                    "nr of plates": "1",
+                                    "blanks": "Sample group 'GROUP_1': On Plate/Chip:  (2x)",
+                                    "plate description": "- Lunatic_test_BE_31JUL2024",
                                     "analyte": "N/A"
                                 }
                             },
@@ -97,7 +109,8 @@
                                         "error feature": "Background (A260)"
                                     }
                                 ]
-                            }
+                            },
+                            "experimental data identifier": "BENCHLING_TEST"
                         }
                     ],
                     "measurement time": "2024-01-08T10:15:59+00:00",
@@ -106,7 +119,8 @@
                         "unit": "#"
                     },
                     "container type": "well plate"
-                }
+                },
+                "analyst": "ImmuneMed"
             }
         ],
         "calculated data aggregate document": {
@@ -210,13 +224,14 @@
             ]
         },
         "data system document": {
-            "ASM file identifier": "Example_Lunatic_Plate_Reader_xlsx_no_header.json",
+            "ASM file identifier": "Example_Lunatic_Plate_Reader_xlsx_with_header.json",
             "data system instance identifier": "N/A",
-            "UNC path": "tests/parsers/unchained_labs_lunatic/testdata/Example_Lunatic_Plate_Reader_xlsx_no_header.xlsx",
-            "file name": "Example_Lunatic_Plate_Reader_xlsx_no_header.xlsx",
+            "UNC path": "tests/parsers/unchained_labs_lunatic_stunner/testdata/Example_Lunatic_Plate_Reader_xlsx_with_header.xlsx",
+            "file name": "Example_Lunatic_Plate_Reader_xlsx_with_header.xlsx",
             "ASM converter name": "allotropy_unchained_labs_lunatic_&_stunner",
             "ASM converter version": "0.1.102",
-            "software name": "Lunatic and Stunner Analysis"
+            "software name": "Lunatic and Stunner Analysis",
+            "software version": "8.2.0.259"
         },
         "device system document": {
             "device identifier": "123456",
