+++ tests/parsers/unchained_labs_lunatic_stunner/testdata/Unchained_Labs_Stunner_example01.json
@@ -9903,7 +9903,7 @@
         "data system document": {
             "ASM file identifier": "Unchained_Labs_Stunner_example01.json",
             "data system instance identifier": "N/A",
-            "UNC path": "tests/parsers/unchained_labs_lunatic/testdata/Unchained_Labs_Stunner_example01.xlsx",
+            "UNC path": "tests/parsers/unchained_labs_lunatic_stunner/testdata/Unchained_Labs_Stunner_example01.xlsx",
             "file name": "Unchained_Labs_Stunner_example01.xlsx",
             "ASM converter name": "allotropy_unchained_labs_lunatic_&_stunner",
             "ASM converter version": "0.1.102",
