+++ tests/parsers/unchained_labs_lunatic_stunner/testdata/spectrum_measurement.csv
