+++ tests/parsers/unchained_labs_lunatic_stunner/testdata/spectrum_measurement.json
@@ -15937,7 +15937,7 @@
         "data system document": {
             "ASM file identifier": "spectrum_measurement.json",
             "data system instance identifier": "N/A",
-            "UNC path": "tests/parsers/unchained_labs_lunatic/testdata/spectrum_measurement.csv",
+            "UNC path": "tests/parsers/unchained_labs_lunatic_stunner/testdata/spectrum_measurement.csv",
             "file name": "spectrum_measurement.csv",
             "ASM converter name": "allotropy_unchained_labs_lunatic_&_stunner",
             "ASM converter version": "0.1.102",
